+++ perf_tests/test_dependencies.py
@@ -11,21 +11,22 @@ from tests.utils import write_test
 
 @pytest.fixture
 def fixture_path_relative(tmpdir_factory):
-    fixture_path = str(tmpdir_factory.mktemp("relative_perf"))
+    fixture_path = str(tmpdir_factory.mktemp("dep_perf_sparse"))
     for module_index in range(10):
         testname = os.path.join(
-            fixture_path, "test_relative_perf{}.py".format(module_index))
+            fixture_path, "test_dep_perf{}.py".format(module_index))
         test_contents = """
 import pytest
 """
         for i in range(40):
             test_contents += """
-@pytest.mark.order(after="test_{}")
+@pytest.mark.dependency(depends=["test_{}"])
 def test_{}():
     assert True
 """.format(i + 50, i)
         for i in range(60):
             test_contents += """
+@pytest.mark.dependency
 def test_{}():
     assert True
 """.format(i + 40)
@@ -35,10 +36,10 @@ def test_{}():
 
 
 @mock.patch("pytest_order.Sorter", TimedSorter)
-def test_performance_relative(fixture_path_relative):
-    """Test performance of after markers that point to tests without
-    an order mark (the usual case)."""
-    args = [fixture_path_relative]
+def test_performance_dependency(fixture_path_relative):
+    """Test performance of dependency markers that point to tests without
+    an order mark (same as for test_relative does for after markers)."""
+    args = ["--order-dependencies", fixture_path_relative]
     TimedSorter.nr_marks = 400
     pytest.main(args, [pytest_order])
-    assert TimedSorter.elapsed < 0.5
+    assert TimedSorter.elapsed < 0.4